requirements.txt
# Generic
hydra-core==1.3.2
tqdm==4.67.1
black==25.1.0
termcolor==3.0.1
seaborn==0.13.2
blobfile==3.0.0
fsspec
git-lfs==1.6
jinja2==3.1.5
pyyaml==6.0.2
torchmetrics==1.7.1

# ML
# NOTE: torch and numpy are NOT pinned here — on Modal the NGC PyTorch
# base image (nvcr.io/nvidia/pytorch:25.02-py3) already provides them
# alongside the matching CUDA / cuDNN / NCCL stack. On CSCS we rely on
# the same NGC image. Reinstalling torch via pip would clobber the
# NGC build with a pure-CPU wheel or a mismatched CUDA version.
einops==0.8.1
fancy-einsum==0.0.3
mauve-text==0.4.0
git+https://github.com/EleutherAI/lm-evaluation-harness
sentencepiece==0.2.0
lightning==2.5.1
pandas==2.2.1
scikit-learn==1.4.0
timm==1.0.15
jupyter==1.1.1
ocifs==1.3.2

# Huggingface
huggingface-hub==0.30.2
safetensors==0.5.2
tokenizers==0.20.3
evaluate==0.4.3
datasets==3.5.0
transformers==4.45.0
accelerate==1.6.0

# Monitoring / configuration
tensorboard==2.16.2
wandb==0.19.9
omegaconf==2.3.0
rich==13.9.4
